+++ git.lua
@@ -1,41 +1,18 @@
 -- preamble: common routines
 
 local path = require('path')
+local git = require('gitutil')
 local matchers = require('matchers')
 local w = require('tables').wrap
 local parser = clink.arg.new_parser
 
 ---
- -- Resolves closest .git directory location.
- -- Navigates subsequently up one level and tries to find .git directory
- -- @param  {string} path Path to directory will be checked. If not provided
- --                       current directory will be used
- -- @return {string} Path to .git directory or nil if such dir not found
-local function get_git_dir(start_dir)
-
-    -- Checks if provided directory contains git directory
-    local function has_git_dir(dir)
-        return #clink.find_dirs(dir..'/.git') > 0
-    end
-
-    -- Set default path to current directory
-    if not start_dir or start_dir == '.' then start_dir = clink.get_cwd() end
-
-    -- If we're already have .git directory here, then return current path
-    if has_git_dir(start_dir) then return start_dir..'/.git' end
-
-    -- Otherwise go up one level and make a recursive call
-    local parent_path =  path.pathname(start_dir)
-    if parent_path ~= start_dir then return get_git_dir(parent_path) end
-end
-
----
  -- Lists remote branches based on packed-refs file from git directory
- -- @param string [git_dir]  Directory where to search file for
+ -- @param string [dir]  Directory where to search file for
  -- @return table  List of remote branches
-local function list_packed_refs(git_dir)
+local function list_packed_refs(dir)
     local result = w()
-    local git_dir = git_dir or get_git_dir()
+    local git_dir = dir or git.get_git_dir()
     if not git_dir then return result end
 
     local packed_refs_file = io.open(git_dir..'/packed-refs')
@@ -53,8 +30,8 @@ local function list_packed_refs(git_dir)
     return result
 end
 
-local function list_remote_branches(git_dir)
-    local git_dir = git_dir or get_git_dir()
+local function list_remote_branches(dir)
+    local git_dir = dir or git.get_git_dir()
     if not git_dir then return w() end
 
     return w(path.list_files(git_dir..'/refs/remotes', '/*',
@@ -66,10 +43,10 @@ end
 ---
  -- Lists local branches for git repo in git_dir directory.
  --
- -- @param string [git_dir]  Git directory, where to search for remote branches
+ -- @param string [dir]  Git directory, where to search for remote branches
  -- @return table  List of branches.
-local function list_local_branches(git_dir)
-    local git_dir = git_dir or get_git_dir()
+local function list_local_branches(dir)
+    local git_dir = dir or git.get_git_dir()
     if not git_dir then return w() end
 
     local result = w(path.list_files(git_dir..'/refs/heads', '/*',
@@ -79,12 +56,12 @@ local function list_local_branches(git_dir)
 end
 
 local branches = function (token)
-    local git_dir = get_git_dir()
+    local git_dir = git.get_git_dir()
     if not git_dir then return w() end
 
     return list_local_branches(git_dir)
-    :filter(function(path)
-        return clink.is_match(token, path)
+    :filter(function(branch)
+        return clink.is_match(token, branch)
     end)
 end
 
@@ -92,19 +69,19 @@ local function alias(token)
     local res = w()
 
     -- Try to resolve .git directory location
-    local git_dir = get_git_dir()
+    local git_dir = git.get_git_dir()
 
     if git_dir == nil then return res end
 
-    f = io.popen("git config --get-regexp alias 2>nul")
+    local f = io.popen("git config --get-regexp alias 2>nul")
     if f == nil then return {} end
 
     for line in f:lines() do
-        local s, e = line:find(" ", 1, true)
-        local alias = line:sub(7, s - 1)
-        local start = alias:find(token, 1, true)
+        local s = line:find(" ", 1, true)
+        local alias_name = line:sub(7, s - 1)
+        local start = alias_name:find(token, 1, true)
         if start and start == 1 then
-            table.insert(res, alias)
+            table.insert(res, alias_name)
         end
     end
 
@@ -113,35 +90,35 @@ local function alias(token)
     return res
 end
 
-local function remotes(token)
-    local remotes = w()
-    local git_dir = get_git_dir()
-    if not git_dir then return remotes end
+local function remotes(token)  -- luacheck: no unused args
+    local result = w()
+    local git_dir = git.get_git_dir()
+    if not git_dir then return result end
 
     local git_config = io.open(git_dir..'/config')
     -- if there is no gitconfig file (WAT?!), return empty list
-    if git_config == nil then return remotes end
+    if git_config == nil then return result end
 
     for line in git_config:lines() do
         local remote = line:match('%[remote "(.*)"%]')
         if (remote) then
-            table.insert(remotes, remote)
+            table.insert(result, remote)
         end
     end
 
     git_config:close()
-    return remotes
+    return result
 end
 
 local function local_or_remote_branches(token)
     -- Try to resolve .git directory location
-    local git_dir = get_git_dir()
+    local git_dir = git.get_git_dir()
     if not git_dir then return w() end
 
     return list_local_branches(git_dir)
     :concat(list_remote_branches(git_dir))
-    :filter(function(path)
-        return clink.is_match(token, path)
+    :filter(function(branch)
+        return clink.is_match(token, branch)
     end)
 end
 
@@ -151,12 +128,12 @@ local function checkout_spec_generator(token)
             return path.is_real_dir(file)
         end)
 
-    local git_dir = get_git_dir()
+    local git_dir = git.get_git_dir()
 
     local local_branches = branches(token)
     local remote_branches = list_remote_branches(git_dir)
-        :filter(function(path)
-            return clink.is_match(token, path)
+        :filter(function(branch)
+            return clink.is_match(token, branch)
         end)
 
     local predicted_branches = list_remote_branches(git_dir)
@@ -178,7 +155,7 @@ local function checkout_spec_generator(token)
     --   * create display filter for completion table to append path separator to each directory entry
     --     since it is not added automatically by readline (see previous point)
     clink.matches_are_files(0)
-    clink.match_display_filter = function (matches)
+    clink.match_display_filter = function ()
         return files:map(function(file)
             return clink.is_dir(file) and file..'\\' or file
         end)
@@ -194,7 +171,7 @@ local function checkout_spec_generator(token)
 end
 
 local function push_branch_spec(token)
-    local git_dir = get_git_dir()
+    local git_dir = git.get_git_dir()
     if not git_dir then return w() end
 
     local plus_prefix = token:sub(0, 1) == '+'
@@ -226,9 +203,10 @@ local function push_branch_spec(token)
         local b = w(clink.find_dirs(git_dir..'/refs/remotes/*'))
         :filter(function(remote) return path.is_real_dir(remote) end)
         :reduce({}, function(result, remote)
-            return w(path.list_files(git_dir..'/refs/remotes/'..remote, '/*', --[[recursive=]]true, --[[reverse_separator=]]true))
-            :filter(function(path)
-                return clink.is_match(remote_branch_spec, path)
+            return w(path.list_files(git_dir..'/refs/remotes/'..remote, '/*',
+                --[[recursive=]]true, --[[reverse_separator=]]true))
+            :filter(function(remote_branch)
+                return clink.is_match(remote_branch_spec, remote_branch)
             end)
             :concat(result)
         end)
@@ -244,9 +222,9 @@ local function push_branch_spec(token)
     end
 end
 
-local stashes = function(token)
+local stashes = function(token)  -- luacheck: no unused args
 
-    local git_dir = get_git_dir()
+    local git_dir = git.get_git_dir()
     if not git_dir then return w() end
 
     local stash_file = io.open(git_dir..'/logs/refs/stash')
@@ -285,7 +263,7 @@ local stashes = function(token)
         table.insert(ret_filter, "stash@{"..(i-1).."}    "..stashes[v])
     end
 
-    clink.match_display_filter = function (matches)
+    clink.match_display_filter = function ()
         return ret_filter
     end
 
@@ -446,9 +424,9 @@ local git_parser = parser(
             "--contains" ,
             "--abbrev",
             "-a", "--all",
-            "-d" .. parser({branches}),
-            "--delete" .. parser({branches}),
-            "-D" .. parser({branches}),
+            "-d" .. parser({branches}):loop(1),
+            "--delete" .. parser({branches}):loop(1),
+            "-D" .. parser({branches}):loop(1),
             "-m", "--move",
             "-M",
             "--list",
@@ -484,7 +462,24 @@ local git_parser = parser(
         ),
         "checkout-index",
         "cherry",
-        "cherry-pick",
+        "cherry-pick"..parser(
+            "-e", "--edit",
+            "-m", "--mainline ",
+            "-n", "--no-commit",
+            "-r",
+            "-x",
+            "--ff",
+             "-s", "-S", "--gpg-sign",
+            "--allow-empty",
+            "--allow-empty-message",
+            "--keep-redundant-commits",
+            "--strategy"..parser({merge_strategies}),
+            "-X"..parser({merge_recursive_options}),
+            "--strategy-option"..parser({merge_recursive_options}),
+            "--continue",
+            "--quit",
+            "--abort"
+        ),
         "citool",
         "clean",
         "clone",
@@ -536,15 +531,23 @@ local git_parser = parser(
         "credential-wincred",
         "daemon",
         "describe",
-        "diff" .. parser({local_or_remote_branches}),
+        "diff" .. parser({local_or_remote_branches, matchers.files}),
         "diff-files",
         "diff-index",
         "diff-tree",
-        "difftool",
+        "difftool"..parser(
+            "-d", "--dir-diff",
+            "-y", "--no-prompt", "--prompt",
+            "-t", "--tool=" -- TODO: complete tool (take from config)
+        ),
         "difftool--helper",
         "fast-export",
         "fast-import",
-        "fetch" .. parser({remotes}),
+        "fetch" .. parser({remotes},
+            "--all",
+            "--prune",
+            "--tags"
+        ),
         "fetch-pack",
         "filter-branch",
         "fmt-merge-msg",
@@ -745,8 +748,8 @@ local git_parser = parser(
         "repo-config",
         "request-pull",
         "rerere",
-        "reset"..parser(
-        -- TODO: Add commit/tree/branch completions
+        -- TODO: Add commit completions
+        "reset"..parser({local_or_remote_branches},
             "-q",
             "-p", "--patch",
             "--soft", "--mixed", "--hard",
@@ -768,7 +771,8 @@ local git_parser = parser(
         "show-ref",
         "stage",
         "stash"..parser({
-            "list", -- TODO: The command takes options applicable to the git log command to control what is shown and how
+            "list", -- TODO: The command takes options applicable to the git log
+                    -- command to control what is shown and how it's done
             "show"..parser({stashes}),
             "drop"..parser({stashes}, "-q", "--quiet"),
             "pop"..parser({stashes}, "--index", "-q", "--quiet"),
@@ -785,38 +789,49 @@ local git_parser = parser(
             }),
         "status",
         "stripspace",
-        "submodule",
+        "submodule"..parser({
+            "add",
+            "init",
+            "deinit",
+            "foreach",
+            "status"..parser("--cached", "--recursive"),
+            "summary",
+            "sync",
+            "update"
+        }, '--quiet'),
         "subtree",
         "svn"..parser({
-		    "init"..parser("-T", "--trunk", "-t", "--tags", "-b", "--branches", "-s", "--stdlayout",
-				"--no-metadata", "--use-svm-props", "--use-svnsync-props", "--rewrite-root",
-				"--rewrite-uuid", "--username", "--prefix"..parser({"origin"}), "--ignore-paths",
-				"--include-paths", "--no-minimize-url"),
-				"fetch"..parser({remotes}, "--localtime", "--parent", "--ignore-paths", "--include-paths",
-				"--log-window-size"),
-				"clone"..parser("-T", "--trunk", "-t", "--tags", "-b", "--branches", "-s", "--stdlayout",
-				"--no-metadata", "--use-svm-props", "--use-svnsync-props", "--rewrite-root",
-				"--rewrite-uuid", "--username", "--prefix"..parser({"origin"}), "--ignore-paths",
-				"--include-paths", "--no-minimize-url", "--preserve-empty-dirs",
-				"--placeholder-filename"),
-				"rebase"..parser({local_or_remote_branches}, {branches}),
-			"dcommit"..parser("--no-rebase", "--commit-url", "--mergeinfo", "--interactive"),
-			"branch"..parser("-m","--message","-t", "--tags", "-d", "--destination", "--username", "--commit-url", "--parents"),
-			"log"..parser("-r", "--revision", "-v", "--verbose", "--limit", "--incremental", "--show-commit", "--oneline"),
-			"find-rev"..parser("--before", "--after"),
-			"reset"..parser("-r", "--revision", "-p", "--parent"),
-			"tag",
-			"blame",
-			"set-tree",
-			"create-ignore",
-			"show-ignore",
-			"mkdirs",
-			"commit-diff",
-			"info",
-			"proplist",
-			"propget",
-			"show-externals",
-			"gc"
+            "init"..parser("-T", "--trunk", "-t", "--tags", "-b", "--branches", "-s", "--stdlayout",
+                "--no-metadata", "--use-svm-props", "--use-svnsync-props", "--rewrite-root",
+                "--rewrite-uuid", "--username", "--prefix"..parser({"origin"}), "--ignore-paths",
+                "--include-paths", "--no-minimize-url"),
+                "fetch"..parser({remotes}, "--localtime", "--parent", "--ignore-paths", "--include-paths",
+                "--log-window-size"),
+                "clone"..parser("-T", "--trunk", "-t", "--tags", "-b", "--branches", "-s", "--stdlayout",
+                "--no-metadata", "--use-svm-props", "--use-svnsync-props", "--rewrite-root",
+                "--rewrite-uuid", "--username", "--prefix"..parser({"origin"}), "--ignore-paths",
+                "--include-paths", "--no-minimize-url", "--preserve-empty-dirs",
+                "--placeholder-filename"),
+                "rebase"..parser({local_or_remote_branches}, {branches}),
+            "dcommit"..parser("--no-rebase", "--commit-url", "--mergeinfo", "--interactive"),
+            "branch"..parser("-m","--message","-t", "--tags", "-d", "--destination",
+                             "--username", "--commit-url", "--parents"),
+            "log"..parser("-r", "--revision", "-v", "--verbose", "--limit",
+                          "--incremental", "--show-commit", "--oneline"),
+            "find-rev"..parser("--before", "--after"),
+            "reset"..parser("-r", "--revision", "-p", "--parent"),
+            "tag",
+            "blame",
+            "set-tree",
+            "create-ignore",
+            "show-ignore",
+            "mkdirs",
+            "commit-diff",
+            "info",
+            "proplist",
+            "propget",
+            "show-externals",
+            "gc"
         }),
         "symbolic-ref",
         "tag",